+++ src/main/java/application/calculei/usecase/Tj6899/dto/CalculateTj6899BetweenDateResponse.java
@@ -3,5 +3,5 @@ package application.calculei.usecase.Tj6899.dto;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record CalculateTj6899BetweenDateResponse(LocalDate dataInicio, LocalDate dataFinal, int diasContados, BigDecimal valorFinal, BigDecimal fatorAcumulado) {
+public record CalculateTj6899BetweenDateResponse(LocalDate dataInicio, LocalDate dataFinal, Long diasContados, BigDecimal valorFinal, BigDecimal fatorAcumulado) {
 }
